Add unit testing with mocha and refactor parseCommand
